requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
livekit==1.3.2
livekit-api==1.3.2
pydub==0.25.1
requests==2.31.0
aiofiles>=24.1.0